refactor(engine): extract daemon tool factories to src/daemon/tools/
