+++ src/daemon/workflow-runner.ts
@@ -32,7 +32,7 @@ import { randomUUID } from 'node:crypto';
 import Anthropic from '@anthropic-ai/sdk';
 import { AnthropicBedrock } from '@anthropic-ai/bedrock-sdk';
 import { AgentLoop } from "./agent-loop.js";
-import type { AgentTool, AgentToolResult, AgentEvent, AgentLoopCallbacks, AgentInternalMessage } from "./agent-loop.js";
+import type { AgentTool, AgentEvent, AgentLoopCallbacks, AgentInternalMessage } from "./agent-loop.js";
 import type { V2ToolContext } from '../mcp/types.js';
 import { executeStartWorkflow } from '../mcp/handlers/v2-execution/start.js';
 import { executeContinueWorkflow } from '../mcp/handlers/v2-execution/index.js';
@@ -54,6 +54,29 @@ import { writeStatsSummary } from './stats-summary.js';
 import { injectPendingSteps } from './turn-end/step-injector.js';
 import { flushConversation } from './turn-end/conversation-flusher.js';
 import { type SessionScope, DefaultFileStateTracker } from './session-scope.js';
+// Tool factories -- extracted to individual files under src/daemon/tools/.
+// Imported for use by constructTools() in this file, and re-exported for backward
+// compatibility (tests and other callers import from workflow-runner.ts).
+import { withWorkrailSession, persistTokens, DAEMON_SESSIONS_DIR } from './tools/_shared.js';
+import { makeContinueWorkflowTool, makeCompleteStepTool } from './tools/continue-workflow.js';
+import { makeBashTool } from './tools/bash.js';
+import { makeReadTool, makeWriteTool, makeEditTool } from './tools/file-tools.js';
+import { makeGlobTool, makeGrepTool } from './tools/glob-grep.js';
+import { makeSpawnAgentTool } from './tools/spawn-agent.js';
+import { makeReportIssueTool } from './tools/report-issue.js';
+import { makeSignalCoordinatorTool } from './tools/signal-coordinator.js';
+// Re-export for backward compatibility (tests and other callers import from workflow-runner.ts).
+export { DAEMON_SESSIONS_DIR, type PersistTokensError } from './tools/_shared.js';
+export { DAEMON_SIGNALS_DIR } from './tools/signal-coordinator.js';
+export {
+  makeContinueWorkflowTool, makeCompleteStepTool,
+  makeBashTool,
+  makeReadTool, makeWriteTool, makeEditTool,
+  makeGlobTool, makeGrepTool,
+  makeSpawnAgentTool,
+  makeReportIssueTool,
+  makeSignalCoordinatorTool,
+};
 
 const execAsync = promisify(exec);
 const execFileAsync = promisify(execFile);
@@ -61,9 +84,6 @@ const execFileAsync = promisify(execFile);
 // ---------------------------------------------------------------------------
 // Constants
 // ---------------------------------------------------------------------------
-
-/** Maximum wall-clock time allowed for a single Bash tool invocation. */
-const BASH_TIMEOUT_MS = 5 * 60 * 1000;
 
 /**
  * Maximum number of prior step notes injected into the session state recap.
@@ -107,28 +127,7 @@ export const DEFAULT_SESSION_TIMEOUT_MINUTES = 30;
  */
 export const DEFAULT_MAX_TURNS = 200;
 
-/**
- * Conditionally spreads workrailSessionId into a daemon event object.
- *
- * WHY: workrailSessionId (the WorkRail sess_xxx ID) is unavailable at session_started
- * time -- it can only be decoded from the continueToken after executeStartWorkflow()
- * returns. All subsequent events include it when available, via this helper.
- * Returns an empty object when sid is null so callers can spread unconditionally.
- */
-function withWorkrailSession(sid: string | null | undefined): { workrailSessionId?: string } {
-  return sid != null ? { workrailSessionId: sid } : {};
-}
-
-/**
- * Directory that holds per-session crash-recovery state files.
- * Each concurrent runWorkflow() call writes to its own <sessionId>.json file,
- * so concurrent sessions never clobber each other.
- *
- * Note: sessionId is a process-local UUID generated at runWorkflow() entry --
- * it is NOT the WorkRail server session ID. The server continueToken is stored
- * as a value inside the file, so crash-resume can retrieve it by reading the file.
- */
-export const DAEMON_SESSIONS_DIR = path.join(os.homedir(), '.workrail', 'daemon-sessions');
+// DAEMON_SESSIONS_DIR is re-exported from './tools/_shared.js' at the top of this file.
 
 /**
  * Maximum age for an orphaned session file before it is treated as definitely stale.
@@ -769,56 +768,8 @@ export interface OrphanedSession {
  *   (recovered sessions use daemon defaults per pitch no-gos).
  *   Old sidecars lacking these fields fall through to discard on the next daemon start.
  */
-/**
- * Error shape for a persistTokens failure.
- * WHY its own type (not a generic string): the code field lets callers
- * distinguish filesystem errors (ENOSPC, EPERM) from logic errors without
- * string parsing.
- */
-export interface PersistTokensError {
-  readonly code: string;
-  readonly message: string;
-}
-
-async function persistTokens(
-  sessionId: string,
-  continueToken: string,
-  checkpointToken: string | null,
-  worktreePath?: string,
-  recoveryContext?: {
-    readonly workflowId: string;
-    readonly goal: string;
-    readonly workspacePath: string;
-  },
-): Promise<Result<void, PersistTokensError>> {
-  try {
-    await fs.mkdir(DAEMON_SESSIONS_DIR, { recursive: true });
-
-    const sessionPath = path.join(DAEMON_SESSIONS_DIR, `${sessionId}.json`);
-    const state = JSON.stringify(
-      {
-        continueToken,
-        checkpointToken,
-        ts: Date.now(),
-        ...(worktreePath !== undefined ? { worktreePath } : {}),
-        ...(recoveryContext !== undefined ? {
-          workflowId: recoveryContext.workflowId,
-          goal: recoveryContext.goal,
-          workspacePath: recoveryContext.workspacePath,
-        } : {}),
-      },
-      null,
-      2,
-    );
-    const tmp = `${sessionPath}.tmp`;
-    await fs.writeFile(tmp, state, 'utf8');
-    await fs.rename(tmp, sessionPath);
-    return ok(undefined);
-  } catch (e: unknown) {
-    const nodeErr = e as NodeJS.ErrnoException;
-    return err({ code: nodeErr.code ?? 'UNKNOWN', message: nodeErr.message ?? String(e) });
-  }
-}
+// PersistTokensError is re-exported from './tools/_shared.js' at the top of this file.
+// persistTokens is imported from './tools/_shared.js' at the top of this file.
 
 /**
  * Append a batch of AgentInternalMessage values to a per-session conversation JSONL file.
@@ -1834,1130 +1785,12 @@ function getSchemas(): Record<string, any> {
   return _schemas;
 }
 
-// ---------------------------------------------------------------------------
-// Tool factories
-// ---------------------------------------------------------------------------
-
-export function makeContinueWorkflowTool(
-  sessionId: string,
-  ctx: V2ToolContext,
-  onAdvance: (nextStepText: string, continueToken: string) => void,
-  onComplete: (notes: string | undefined, artifacts?: readonly unknown[]) => void,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  schemas: Record<string, any>,
-  // Optional injection point for testing -- defaults to the real implementation.
-  _executeContinueWorkflowFn: typeof executeContinueWorkflow = executeContinueWorkflow,
-  emitter?: DaemonEventEmitter,
-  workrailSessionId?: string | null,
-): AgentTool {
-  return {
-    name: 'continue_workflow',
-    description:
-      '[DEPRECATED in daemon sessions -- use complete_step instead] ' +
-      'Advance the WorkRail workflow to the next step. Call this after completing all work ' +
-      'required by the current step. Include your notes in notesMarkdown. ' +
-      'When the step requires an assessment gate, include wr.assessment objects in artifacts.',
-    inputSchema: schemas['ContinueWorkflowParams'],
-    label: 'Continue Workflow',
-
-    execute: async (
-      _toolCallId: string,
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      console.log(`[WorkflowRunner] Tool: continue_workflow sessionId=${sessionId}`);
-      emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'continue_workflow', summary: (params.intent as string | undefined) ?? 'advance', ...withWorkrailSession(workrailSessionId) });
-      const result = await _executeContinueWorkflowFn(
-        {
-          continueToken: params.continueToken,
-          intent: (params.intent ?? 'advance') as 'advance' | 'rehydrate',
-          // WHY: output is constructed when either notesMarkdown or artifacts is present.
-          // Agents may need to submit assessment artifacts without notes (e.g. when the
-          // step's only requirement is an assessment gate). Using `?.length` prevents an
-          // empty artifacts array from constructing a spurious output object.
-          output: (params.notesMarkdown || (params.artifacts as unknown[] | undefined)?.length)
-            ? {
-                ...(params.notesMarkdown ? { notesMarkdown: params.notesMarkdown } : {}),
-                ...(params.artifacts ? { artifacts: params.artifacts } : {}),
-              }
-            : undefined,
-          context: params.context,
-        },
-        ctx,
-      );
-
-      if (result.isErr()) {
-        throw new Error(`continue_workflow failed: ${result.error.kind} -- ${JSON.stringify(result.error)}`);
-      }
-
-      const out = result.value.response;
-
-      // Persist tokens atomically before returning -- crash safety invariant.
-      // WHY continueToken vs retryToken: for a blocked response, nextCall.params.continueToken
-      // is the retry token (retryContinueToken for retryable, or continueToken for non-retryable).
-      // Persisting this ensures crash recovery resumes with the correct token.
-      const continueToken = out.continueToken ?? '';
-      const checkpointToken = out.checkpointToken ?? null;
-      const persistToken = (out.kind === 'blocked' ? out.nextCall?.params.continueToken : undefined) ?? continueToken;
-      if (persistToken) {
-        const persistResult = await persistTokens(sessionId, persistToken, checkpointToken);
-        // WHY log-and-continue (not throw): a persist failure degrades crash recovery but
-        // the session is still live and the LLM has the token in memory. Killing the session
-        // here loses in-progress work. Invariant 4.3: onAdvance/onTokenUpdate must still fire.
-        if (persistResult.kind === 'err') {
-          console.warn(`[WorkflowRunner] persistTokens failed (continue_workflow): ${persistResult.error.code} -- ${persistResult.error.message}`);
-        }
-      }
-
-      // WHY: when the engine returns a blocked response, the step did NOT advance.
-      // Calling onAdvance() would erroneously signal advancement and cause the agent
-      // to loop forever believing it moved forward. Return feedback instead so the
-      // agent knows what to fix and can retry the same step.
-      if (out.kind === 'blocked') {
-        const retryToken = out.nextCall?.params.continueToken ?? continueToken;
-        const lines: string[] = ['## Step blocked -- action required\n'];
-
-        for (const blocker of out.blockers.blockers) {
-          lines.push(blocker.message);
-          if (blocker.suggestedFix) {
-            lines.push(`\nWhat to do: ${blocker.suggestedFix}`);
-          }
-          lines.push('');
-        }
-
-        if (out.validation) {
-          if (out.validation.issues.length > 0) {
-            lines.push('**Issues:**');
-            for (const issue of out.validation.issues) lines.push(`- ${issue}`);
-            lines.push('');
-          }
-          if (out.validation.suggestions.length > 0) {
-            lines.push('**Suggestions:**');
-            for (const s of out.validation.suggestions) lines.push(`- ${s}`);
-            lines.push('');
-          }
-        }
-
-        if (out.assessmentFollowup) {
-          lines.push(`**Follow-up required:** ${out.assessmentFollowup.title}`);
-          lines.push(out.assessmentFollowup.guidance);
-          lines.push('');
-        }
-
-        if (out.retryable) {
-          lines.push(`Retry the same step with corrected output.\n\ncontinueToken: ${retryToken}`);
-        } else {
-          lines.push(`You cannot proceed without resolving this. Inform the user and wait for their response, then call continue_workflow.\n\ncontinueToken: ${retryToken}`);
-        }
-
-        const feedback = lines.join('\n');
-        return {
-          content: [{ type: 'text', text: feedback }],
-          details: out,
-        };
-      }
-
-      if (out.isComplete) {
-        // Pass the agent's notes and artifacts from this final step to onComplete so the
-        // trigger layer can extract the structured handoff artifact for delivery, and so
-        // coordinators can read typed artifacts via WorkflowRunSuccess.lastStepArtifacts.
-        onComplete(
-          params.notesMarkdown as string | undefined,
-          Array.isArray(params.artifacts) ? (params.artifacts as readonly unknown[]) : undefined,
-        );
-        return {
-          content: [{ type: 'text', text: 'Workflow complete. All steps have been executed.' }],
-          details: out,
-        };
-      }
-
-      const pending = out.pending;
-      const stepText = pending
-        ? `## Next step: ${pending.title}\n\n${pending.prompt}\n\ncontinueToken: ${continueToken}`
-        : `Step advanced. continueToken: ${continueToken}`;
-
-      onAdvance(stepText, continueToken);
-
-      return {
-        content: [{ type: 'text', text: stepText }],
-        details: out,
-      };
-    },
-  };
-}
-
-/**
- * Build the complete_step tool for daemon sessions.
- *
- * WHY this tool exists: continue_workflow requires the LLM to round-trip a
- * continueToken (an HMAC-signed opaque token). The LLM frequently mangles
- * this token, causing TOKEN_BAD_SIGNATURE errors that kill sessions. complete_step
- * eliminates this failure mode by having the daemon inject the continueToken
- * internally -- the LLM only provides notes, artifacts, and context.
- *
- * WHY two token-update paths: the continueToken must be updated on both
- * (a) successful advance: getCurrentToken() returns the new next-step token
- *     from the response, which onAdvance will have stored before the next call.
- * (b) blocked retry: the engine returns a retryContinueToken that must be used
- *     on the retry call; onTokenUpdate updates the closure variable so the next
- *     complete_step call injects the correct retry token.
- * Both paths are mutually exclusive (kind: 'ok' vs kind: 'blocked') and cannot
- * race because AgentLoop runs tools sequentially (toolExecution: 'sequential').
- *
- * WHY getCurrentToken is a getter (not a value): the closure variable
- * currentContinueToken in runWorkflow() is updated after each step advance.
- * The getter captures the variable by reference so each complete_step call
- * reads the current token at call time, not at construction time.
- *
- * @param sessionId - Process-local UUID for crash-recovery token persistence.
- * @param ctx - V2ToolContext from the shared DI container.
- * @param getCurrentToken - Getter that returns the current continueToken from the
- *   runWorkflow() closure. Called at tool execution time, not construction time.
- * @param onAdvance - Called after a successful step advance with the next step text
- *   and the new continueToken. Appends step text to pendingSteerParts and updates currentContinueToken.
- * @param onComplete - Called when the workflow is complete.
- * @param onTokenUpdate - Called when the continueToken changes without an advance
- *   (i.e., on a blocked retry). Updates currentContinueToken in the runWorkflow() closure.
- * @param schemas - Plain JSON Schema map from getSchemas().
- * @param _executeContinueWorkflowFn - Optional injection point for testing.
- * @param emitter - Optional event emitter for structured lifecycle events.
- * @param workrailSessionId - WorkRail session ID for event correlation.
- */
-export function makeCompleteStepTool(
-  sessionId: string,
-  ctx: V2ToolContext,
-  getCurrentToken: () => string,
-  onAdvance: (nextStepText: string, continueToken: string) => void,
-  onComplete: (notes: string | undefined, artifacts?: readonly unknown[]) => void,
-  onTokenUpdate: (t: string) => void,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  schemas: Record<string, any>,
-  // Optional injection point for testing -- defaults to the real implementation.
-  _executeContinueWorkflowFn: typeof executeContinueWorkflow = executeContinueWorkflow,
-  emitter?: DaemonEventEmitter,
-  workrailSessionId?: string | null,
-): AgentTool {
-  return {
-    name: 'complete_step',
-    description:
-      'Mark the current WorkRail workflow step as complete and advance to the next one. ' +
-      'Call this after completing all work required by the current step. ' +
-      'Include your substantive notes (min 50 characters) describing what you did. ' +
-      'The daemon manages the session token internally -- you do not need a continueToken. ' +
-      'When the step requires an assessment gate, include wr.assessment objects in artifacts.',
-    inputSchema: schemas['CompleteStepParams'],
-    label: 'Complete Step',
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    execute: async (
-      _toolCallId: string,
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      console.log(`[WorkflowRunner] Tool: complete_step sessionId=${sessionId}`);
-      emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'complete_step', summary: 'advance', ...withWorkrailSession(workrailSessionId) });
-
-      // WHY runtime validation: JSON Schema minLength is informational to the LLM
-      // but NOT enforced by AgentLoop. We must validate here so the LLM gets a
-      // clear error immediately, rather than a downstream blocked response from
-      // the engine. Fail fast at the boundary.
-      const notes = params.notes as string | undefined;
-      if (!notes || notes.length < 50) {
-        throw new Error(
-          `complete_step: notes is required and must be at least 50 characters. ` +
-          `Provide substantive notes describing what you did, what you produced, and any notable decisions. ` +
-          `Current length: ${notes?.length ?? 0} characters.`,
-        );
-      }
-
-      // WHY inject getCurrentToken(): the daemon holds the continueToken in a
-      // closure variable (currentContinueToken in runWorkflow()). The LLM never
-      // sees this token -- we inject it here so the engine can authenticate the
-      // advance call. This is the core value of complete_step over continue_workflow.
-      const continueToken = getCurrentToken();
-
-      const result = await _executeContinueWorkflowFn(
-        {
-          continueToken,
-          intent: 'advance',
-          // WHY: output is constructed when notes is present (always true after validation)
-          // or when artifacts is a non-empty array (e.g. assessment-only steps without notes,
-          // though complete_step always requires notes). An empty artifacts array must not
-          // spread {} or {} with artifacts: [] -- use ?.length to guard against this.
-          output: (notes || (params.artifacts as unknown[] | undefined)?.length)
-            ? {
-                notesMarkdown: notes,
-                ...((params.artifacts as unknown[] | undefined)?.length ? { artifacts: params.artifacts } : {}),
-              }
-            : undefined,
-          context: params.context,
-        },
-        ctx,
-      );
-
-      if (result.isErr()) {
-        throw new Error(`complete_step failed: ${result.error.kind} -- ${JSON.stringify(result.error)}`);
-      }
-
-      const out = result.value.response;
-
-      // Persist tokens atomically before returning -- crash safety invariant.
-      // WHY this must happen before onAdvance/onTokenUpdate: a crash between
-      // executeContinueWorkflow returning and the token being persisted would
-      // leave no recoverable state. Persisting first ensures crash recovery works.
-      const newContinueToken = out.continueToken ?? '';
-      const checkpointToken = out.checkpointToken ?? null;
-      // WHY blocked uses retry token: on a blocked response, the engine returns a
-      // retryContinueToken (via nextCall.params.continueToken). The session token
-      // advances to this retry token -- the original session token is consumed.
-      const persistToken = (out.kind === 'blocked' ? out.nextCall?.params.continueToken : undefined) ?? newContinueToken;
-      if (persistToken) {
-        const persistResult = await persistTokens(sessionId, persistToken, checkpointToken);
-        // WHY log-and-continue (not throw): a persist failure degrades crash recovery but
-        // the session is still live. Invariant 4.3: onAdvance/onTokenUpdate must still fire.
-        if (persistResult.kind === 'err') {
-          console.warn(`[WorkflowRunner] persistTokens failed (complete_step): ${persistResult.error.code} -- ${persistResult.error.message}`);
-        }
-      }
-
-      // WHY onTokenUpdate on blocked: the next complete_step call must inject the
-      // retry token (not the original session token). We update the closure variable
-      // so getCurrentToken() returns the correct retry token on the next call.
-      // This is a separate path from onAdvance because a blocked response does NOT
-      // advance the step -- it only changes which token is valid for retry.
-      if (out.kind === 'blocked') {
-        const retryToken = out.nextCall?.params.continueToken ?? newContinueToken;
-        // Update the closure token to the retry token for the next complete_step call.
-        onTokenUpdate(retryToken);
-
-        const lines: string[] = ['## Step blocked -- action required\n'];
-
-        for (const blocker of out.blockers.blockers) {
-          lines.push(blocker.message);
-          if (blocker.suggestedFix) {
-            lines.push(`\nWhat to do: ${blocker.suggestedFix}`);
-          }
-          lines.push('');
-        }
-
-        if (out.validation) {
-          if (out.validation.issues.length > 0) {
-            lines.push('**Issues:**');
-            for (const issue of out.validation.issues) lines.push(`- ${issue}`);
-            lines.push('');
-          }
-          if (out.validation.suggestions.length > 0) {
-            lines.push('**Suggestions:**');
-            for (const s of out.validation.suggestions) lines.push(`- ${s}`);
-            lines.push('');
-          }
-        }
-
-        if (out.assessmentFollowup) {
-          lines.push(`**Follow-up required:** ${out.assessmentFollowup.title}`);
-          lines.push(out.assessmentFollowup.guidance);
-          lines.push('');
-        }
-
-        if (out.retryable) {
-          lines.push(`Retry the same step: call complete_step again with corrected notes.`);
-        } else {
-          lines.push(`You cannot proceed without resolving this. Inform the user and wait for their response, then call complete_step.`);
-        }
-
-        const feedback = lines.join('\n');
-        return {
-          content: [{ type: 'text', text: feedback }],
-          details: out,
-        };
-      }
-
-      if (out.isComplete) {
-        // Forward artifacts alongside notes so WorkflowRunSuccess.lastStepArtifacts is
-        // populated for coordinator consumption. See docs/discovery/artifacts-coordinator-channel.md.
-        onComplete(notes, Array.isArray(params.artifacts) ? (params.artifacts as readonly unknown[]) : undefined);
-        return {
-          content: [{ type: 'text', text: JSON.stringify({ status: 'complete' }) }],
-          details: out,
-        };
-      }
-
-      const pending = out.pending;
-      // WHY no continueToken in the response text: the LLM does not need the token.
-      // Including it would invite the LLM to store it and pass it to continue_workflow,
-      // defeating the purpose of complete_step.
-      const nextStepTitle = pending?.title ?? 'Next step';
-      const stepText = pending
-        ? `${JSON.stringify({ status: 'advanced', nextStep: pending.title })}\n\n## ${pending.title}\n\n${pending.prompt}`
-        : JSON.stringify({ status: 'advanced', nextStep: nextStepTitle });
-
-      onAdvance(stepText, newContinueToken);
-
-      return {
-        content: [{ type: 'text', text: stepText }],
-        details: out,
-      };
-    },
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeBashTool(workspacePath: string, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Bash',
-    description:
-      'Execute a shell command. Throws on failure (non-zero exit with stderr, or exit code 2+). ' +
-      'Exit code 1 with empty stderr is treated as "no match found" (standard grep semantics) and ' +
-      'returns empty output without throwing. ' +
-      `Maximum execution time: ${BASH_TIMEOUT_MS / 1000}s.`,
-    inputSchema: schemas['BashParams'],
-    label: 'Bash',
-
-    execute: async (
-      _toolCallId: string,
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.command !== 'string' || !params.command) throw new Error('Bash: command must be a non-empty string');
-      console.log(`[WorkflowRunner] Tool: bash "${String(params.command).slice(0, 80)}"`);
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Bash', summary: String(params.command).slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-      const cwd = params.cwd ?? workspacePath;
-      try {
-        const { stdout, stderr } = await execAsync(params.command, {
-          cwd,
-          timeout: BASH_TIMEOUT_MS,
-          shell: '/bin/bash',
-        });
-        const output = [stdout, stderr].filter(Boolean).join('\n');
-        return {
-          content: [{ type: 'text', text: output || '(no output)' }],
-          details: { stdout, stderr },
-        };
-      } catch (err: unknown) {
-        // Node's child_process.exec errors (ExecException) attach stdout, stderr,
-        // code, and signal as own properties. Extract them so the agent can reason
-        // about what went wrong and retry intelligently.
-        // WHY rawCode: `code` is null (not a number) when the process was killed by
-        // a signal (e.g. SIGTERM on exec timeout). Check rawCode before coalescing
-        // so the signal branch is reachable -- `rawCode ?? 'unknown'` would collapse
-        // null to 'unknown' and make signal unreachable.
-        const e = err as { stdout?: unknown; stderr?: unknown; code?: unknown; signal?: unknown };
-        const stdout = String(e.stdout ?? '');
-        const stderr = String(e.stderr ?? '');
-        const rawCode = e.code;
-        const signal = e.signal;
-
-        // Exit code 1 with empty stderr is "no match found" -- not a real error.
-        // This is standard POSIX semantics for grep (exit 1 = no lines matched,
-        // exit 2 = genuine error). Applying the same rule to any command that exits
-        // 1 with no error output is safe: a command that fails silently is almost
-        // always reporting "nothing found", not a broken execution.
-        // WHY `rawCode === 1 && !stderr.trim()`: rawCode is a number when set by
-        // normal process exit; the trim() guard ignores purely-whitespace stderr.
-        // This branch is NOT entered for exit 2+ (always an error) or when stderr
-        // is non-empty (the process wrote a diagnostic message, so it is a real error).
-        if (rawCode === 1 && !stderr.trim()) {
-          return {
-            content: [{ type: 'text', text: stdout || '(no output)' }],
-            details: { stdout, stderr },
-          };
-        }
-
-        const exitInfo = rawCode != null
-          ? `exit ${String(rawCode)}`
-          : signal
-            ? `signal ${String(signal)}`
-            : 'exit unknown';
-        throw new Error(
-          `Command failed: ${params.command} (${exitInfo})\nSTDOUT:\n${stdout}\nSTDERR:\n${stderr}`,
-        );
-      }
-    },
-  };
-}
+// Tool factories are implemented in src/daemon/tools/ and re-exported at the top of this file.
 
 // ---------------------------------------------------------------------------
-// File navigation and editing tools
+// writeStuckOutboxEntry -- stuck detection helper
 // ---------------------------------------------------------------------------
 
-/**
- * Normalizes LLM-generated typographic quotes to their ASCII equivalents.
- *
- * WHY: LLMs trained on typographic text often generate curly single quotes (\u2018, \u2019)
- * and curly double quotes (\u201C, \u201D) where the user intended straight ASCII quotes.
- * When these appear in an `old_string` for Edit, the exact match fails. This function
- * tries the normalized form as a fallback, recovering from the most common LLM quote variants.
- *
- * Handles ~95% of LLM quote variants. Full Unicode normalization is YAGNI (YELLOW-2).
- */
-function findActualString(fileContent: string, oldString: string): string | null {
-  if (fileContent.includes(oldString)) return oldString;
-  // Handle LLM-generated typographic quotes
-  const normalized = oldString
-    .replace(/[\u2018\u2019]/g, "'")
-    .replace(/[\u201C\u201D]/g, '"')
-    .replace(/\u2013/g, '-')
-    .replace(/\u2014/g, '--');
-  if (fileContent.includes(normalized)) return normalized;
-  return null;
-}
-
-/** Max file size Read will return without offset/limit (256 KiB). */
-const READ_SIZE_CAP_BYTES = 256 * 1024;
-
-/** Glob paths always excluded from makeGlobTool results. */
-const GLOB_ALWAYS_EXCLUDE = ['**/node_modules/**', '**/.git/**', '**/dist/**', '**/build/**'];
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeReadTool(readFileState: Map<string, ReadFileState>, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Read',
-    description:
-      'Read the contents of a file at the given absolute path. ' +
-      'Content is returned in cat -n format: each line is prefixed with its 1-indexed line number and a tab character (e.g. "1\\tline one\\n2\\tline two"). ' +
-      'Use offset (0-indexed start line) and limit (max lines) to read a slice of a large file.',
-    inputSchema: schemas['ReadParams'],
-    label: 'Read',
-
-    execute: async (
-      _toolCallId: string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.filePath !== 'string' || !params.filePath) throw new Error('Read: filePath must be a non-empty string');
-      const filePath: string = params.filePath;
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Read', summary: filePath.slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-
-      // Block device paths to prevent reads from infinite streams
-      const devPaths = ['/dev/stdin', '/dev/tty', '/dev/zero', '/dev/random', '/dev/full', '/dev/urandom'];
-      if (devPaths.some(d => filePath === d)) {
-        throw new Error(`Refusing to read device path: ${filePath}`);
-      }
-
-      const stat = await fs.stat(filePath);
-      const offset: number = params.offset ?? 0;
-      const limit: number | undefined = params.limit;
-      const isPaginated = params.offset !== undefined || params.limit !== undefined;
-      if (!isPaginated && stat.size > READ_SIZE_CAP_BYTES) {
-        throw new Error(
-          `File is too large to read at once (${stat.size} bytes, cap is ${READ_SIZE_CAP_BYTES} bytes). ` +
-          `Use offset and limit parameters to read a specific range of lines.`,
-        );
-      }
-
-      const rawContent = await fs.readFile(filePath, 'utf8');
-      const allLines = rawContent.split('\n');
-      const isPartialView = offset !== 0 || limit != null;
-
-      const slicedLines = limit != null ? allLines.slice(offset, offset + limit) : allLines.slice(offset);
-      const startLine = offset; // 0-indexed offset -> 1-indexed line numbers start at offset+1
-      const formatted = slicedLines.map((l, i) => `${startLine + i + 1}\t${l}`).join('\n');
-
-      // Store in readFileState so Edit and Write can enforce staleness checks
-      readFileState.set(filePath, { content: rawContent, timestamp: stat.mtimeMs, isPartialView });
-
-      return {
-        content: [{ type: 'text', text: formatted }],
-        details: { filePath, totalLines: allLines.length, returnedLines: slicedLines.length, offset, isPartialView },
-      };
-    },
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeWriteTool(readFileState: Map<string, ReadFileState>, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Write',
-    description:
-      'Write content to a file at the given absolute path. Creates parent directories if needed. ' +
-      'For existing files: the file must have been read in this session and must not have changed on disk since then. ' +
-      'For new files (path does not exist): no prior read is required.',
-    inputSchema: schemas['WriteParams'],
-    label: 'Write',
-
-    execute: async (
-      _toolCallId: string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.filePath !== 'string' || !params.filePath) throw new Error('Write: filePath must be a non-empty string');
-      if (typeof params.content !== 'string') throw new Error('Write: content must be a string');
-      const filePath: string = params.filePath;
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Write', summary: filePath.slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-
-      // Staleness guard: only for existing files. New files bypass the check entirely.
-      // WHY: a new file has never been read, so there is nothing stale to guard against.
-      let existsOnDisk = false;
-      try {
-        await fs.access(filePath);
-        existsOnDisk = true;
-      } catch {
-        // File does not exist -- new file, no guard needed
-      }
-
-      if (existsOnDisk) {
-        const state = readFileState.get(filePath);
-        if (!state) {
-          throw new Error(
-            `File has not been read in this session. Call Read first before writing to it: ${filePath}`,
-          );
-        }
-        const stat = await fs.stat(filePath);
-        if (stat.mtimeMs !== state.timestamp) {
-          throw new Error(
-            `File has been modified since it was read. Re-read before writing: ${filePath}`,
-          );
-        }
-      }
-
-      await fs.mkdir(path.dirname(filePath), { recursive: true });
-      await fs.writeFile(filePath, params.content, 'utf8');
-
-      // Update readFileState so a subsequent Edit/Write sees the new mtime
-      const newStat = await fs.stat(filePath);
-      readFileState.set(filePath, { content: params.content, timestamp: newStat.mtimeMs, isPartialView: false });
-
-      return {
-        content: [{ type: 'text', text: `Written ${params.content.length} bytes to ${filePath}` }],
-        details: { filePath, length: params.content.length },
-      };
-    },
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeGlobTool(workspacePath: string, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Glob',
-    description:
-      'Find files matching a glob pattern. Returns newline-separated relative file paths, sorted by modification time descending. ' +
-      'node_modules, .git, dist, and build directories are always excluded. ' +
-      'Results are capped at 100 files.',
-    inputSchema: schemas['GlobParams'],
-    label: 'Glob',
-
-    execute: async (
-      _toolCallId: string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.pattern !== 'string' || !params.pattern) throw new Error('Glob: pattern must be a non-empty string');
-      const pattern: string = params.pattern;
-      const searchRoot: string = params.path ?? workspacePath;
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Glob', summary: pattern.slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-
-      const GLOB_LIMIT = 100;
-
-      let paths: string[];
-      try {
-        paths = await tinyGlob(pattern, {
-          cwd: searchRoot,
-          ignore: GLOB_ALWAYS_EXCLUDE,
-          absolute: false,
-        });
-      } catch {
-        // Non-existent search root or invalid pattern returns empty
-        paths = [];
-      }
-
-      // Sort by mtime descending (most recently modified first)
-      const withMtimes = await Promise.all(
-        paths.map(async (p) => {
-          try {
-            const stat = await fs.stat(path.join(searchRoot, p));
-            return { p, mtime: stat.mtimeMs };
-          } catch {
-            return { p, mtime: 0 };
-          }
-        }),
-      );
-      withMtimes.sort((a, b) => b.mtime - a.mtime);
-
-      const sorted = withMtimes.map(x => x.p);
-      const truncated = sorted.length > GLOB_LIMIT;
-      const result = sorted.slice(0, GLOB_LIMIT);
-
-      let text = result.join('\n');
-      if (truncated) {
-        text += '\n[Results truncated at 100 files]';
-      }
-
-      return {
-        content: [{ type: 'text', text: text || '(no matches)' }],
-        details: { pattern, searchRoot, matchCount: sorted.length, truncated },
-      };
-    },
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeGrepTool(workspacePath: string, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Grep',
-    description:
-      'Search file contents using ripgrep (rg). Fast regex search with optional context lines, file-type filtering, and case-insensitive mode. ' +
-      'output_mode: "files_with_matches" (default) returns only file paths; "content" returns matching lines; "count" returns match counts per file. ' +
-      'node_modules and .git are always excluded.',
-    inputSchema: schemas['GrepParams'],
-    label: 'Grep',
-
-    execute: async (
-      _toolCallId: string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.pattern !== 'string' || !params.pattern) throw new Error('Grep: pattern must be a non-empty string');
-      const pattern: string = params.pattern;
-      const searchPath: string = params.path ?? workspacePath;
-      const outputMode: string = params.output_mode ?? 'files_with_matches';
-      const headLimit: number = params.head_limit ?? 250;
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Grep', summary: pattern.slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-
-      const args: string[] = [
-        '--hidden',
-        '--glob', '!node_modules',
-        '--glob', '!.git',
-        '--max-columns', '500',
-      ];
-
-      if (params['-i']) args.push('-i');
-      if (params.glob) { args.push('--glob', params.glob); }
-      if (params.type) { args.push('--type', params.type); }
-
-      switch (outputMode) {
-        case 'files_with_matches':
-          args.push('--files-with-matches');
-          break;
-        case 'count':
-          args.push('--count');
-          break;
-        case 'content':
-          args.push('--vimgrep');
-          if (params.context != null) { args.push('-C', String(params.context)); }
-          break;
-      }
-
-      args.push('--', pattern, searchPath);
-
-      let stdout: string;
-      try {
-        const result = await execFileAsync('rg', args, { cwd: workspacePath, maxBuffer: 10 * 1024 * 1024 });
-        stdout = result.stdout;
-      } catch (err: unknown) {
-        const nodeErr = err as NodeJS.ErrnoException & { stdout?: string; stderr?: string; code?: number | string };
-        // ENOENT: rg binary not found
-        if (nodeErr.code === 'ENOENT') {
-          throw new Error(
-            'ripgrep (rg) is not installed. Install it with: brew install ripgrep (macOS) or apt install ripgrep (Ubuntu/Debian).',
-          );
-        }
-        // exit code 1 from rg means "no matches found" -- not an error
-        if (typeof nodeErr.code === 'number' && nodeErr.code === 1) {
-          return {
-            content: [{ type: 'text', text: '(no matches)' }],
-            details: { pattern, searchPath, outputMode },
-          };
-        }
-        throw new Error(`rg failed: ${nodeErr.message ?? String(err)}`);
-      }
-
-      // Apply head_limit
-      const lines = stdout.split('\n').filter(l => l.length > 0);
-      const truncated = lines.length > headLimit;
-      let result = lines.slice(0, headLimit).join('\n');
-      if (truncated) {
-        result += `\n[Results truncated at ${headLimit} lines. Use a more specific pattern or increase head_limit.]`;
-      }
-
-      return {
-        content: [{ type: 'text', text: result || '(no matches)' }],
-        details: { pattern, searchPath, outputMode, lineCount: lines.length, truncated },
-      };
-    },
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function makeEditTool(workspacePath: string, readFileState: Map<string, ReadFileState>, schemas: Record<string, any>, sessionId?: string, emitter?: DaemonEventEmitter, workrailSessionId?: string | null): AgentTool {
-  return {
-    name: 'Edit',
-    description:
-      'Perform an exact string replacement in a file. ' +
-      'The file must have been read in this session via the Read tool. ' +
-      'By default, old_string must appear exactly once; use replace_all=true to replace all occurrences. ' +
-      'Do NOT include line-number prefixes (e.g. "1\\t") from Read output in old_string or new_string.',
-    inputSchema: schemas['EditParams'],
-    label: 'Edit',
-
-    execute: async (
-      _toolCallId: string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      params: any,
-    ): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.file_path !== 'string' || !params.file_path) throw new Error('Edit: file_path must be a non-empty string');
-      if (typeof params.old_string !== 'string') throw new Error('Edit: old_string must be a string');
-      if (typeof params.new_string !== 'string') throw new Error('Edit: new_string must be a string');
-      const rawFilePath: string = params.file_path;
-      const absoluteFilePath = path.isAbsolute(rawFilePath)
-        ? rawFilePath
-        : path.join(workspacePath, rawFilePath);
-      // Enforce workspace boundary: prevent edits outside the workspace
-      if (!absoluteFilePath.startsWith(workspacePath)) {
-        throw new Error(`Edit target is outside the workspace: ${rawFilePath}`);
-      }
-      const filePath: string = absoluteFilePath;
-      const oldString: string = params.old_string;
-      const newString: string = params.new_string;
-      const replaceAll: boolean = params.replace_all ?? false;
-
-      if (sessionId) emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'Edit', summary: filePath.slice(0, 80), ...withWorkrailSession(workrailSessionId) });
-
-      // Validate that old_string and new_string differ
-      if (oldString === newString) {
-        throw new Error('old_string and new_string are identical. No edit needed.');
-      }
-
-      // Read-before-write enforcement
-      const state = readFileState.get(filePath);
-      if (!state) {
-        throw new Error(
-          `File has not been read in this session. Call Read first before editing: ${filePath}`,
-        );
-      }
-
-      // Staleness check
-      let stat: Awaited<ReturnType<typeof fs.stat>>;
-      try {
-        stat = await fs.stat(filePath);
-      } catch {
-        throw new Error(`File not found: ${filePath}. It may have been deleted after it was read.`);
-      }
-      if (stat.mtimeMs !== state.timestamp) {
-        throw new Error(
-          `File has been modified since it was read. Re-read before editing: ${filePath}`,
-        );
-      }
-
-      const currentContent = await fs.readFile(filePath, 'utf8');
-
-      // Find the actual string to replace (with optional curly-quote normalization)
-      const actualString = findActualString(currentContent, oldString);
-      if (actualString === null) {
-        throw new Error(
-          `String to replace not found in file. Make sure old_string exactly matches the file content ` +
-          `(do not include line-number prefixes from Read output): ${filePath}`,
-        );
-      }
-
-      // Count occurrences
-      const occurrences = currentContent.split(actualString).length - 1;
-      if (!replaceAll && occurrences > 1) {
-        throw new Error(
-          `old_string appears ${occurrences} times in the file. ` +
-          `Provide a more specific string that matches exactly once, or set replace_all=true to replace all occurrences.`,
-        );
-      }
-
-      // Apply the replacement
-      const updatedContent = replaceAll
-        ? currentContent.split(actualString).join(newString)
-        : currentContent.replace(actualString, newString);
-
-      await fs.writeFile(filePath, updatedContent, 'utf8');
-
-      // Update readFileState with new content and new mtime
-      const newStat = await fs.stat(filePath);
-      readFileState.set(filePath, { content: updatedContent, timestamp: newStat.mtimeMs, isPartialView: false });
-
-      return {
-        content: [{ type: 'text', text: `The file ${filePath} has been updated successfully.` }],
-        details: { filePath, occurrencesReplaced: occurrences },
-      };
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// spawn_agent tool
-// ---------------------------------------------------------------------------
-
-/**
- * Build the spawn_agent tool for daemon sessions.
- *
- * Spawns a child WorkRail session in-process, blocking the parent until the child completes.
- * Returns `{ childSessionId, outcome, notes }` as a JSON string so the parent LLM can reason
- * about the child's result and decide how to proceed.
- *
- * WHY in-process (not via TriggerRouter.dispatch): dispatch() uses a global Semaphore and is
- * fire-and-forget. Calling it from inside a running session would deadlock (parent holds a slot
- * waiting for child to acquire one). Direct runWorkflow() call is naturally blocking via await.
- *
- * WHY pre-create session (Candidate 2): calls executeStartWorkflow() first so childSessionId is
- * deterministic and the child is observable in the session store from the moment execute() is called.
- * Crash-before-start is observable (zombie session with parentSessionId). Adapts the proven
- * _preAllocatedStartResponse pattern from console-routes.ts.
- *
- * WHY errors as data (not throw): a child failure should not abort the parent session. The parent
- * LLM receives { outcome: 'error', notes: errorMessage } and decides how to proceed. This is
- * different from other tools (complete_step, Bash) which throw on failure because those failures
- * ARE errors in the parent session's own execution.
- *
- * WHY currentDepth is a constructor parameter (not read from store): depth is set at factory
- * construction time by runWorkflow(), which reads trigger.spawnDepth. For checkpoint-resumed
- * sessions, the daemon restarts the AgentLoop from scratch and re-reads trigger.spawnDepth --
- * the constructor parameter is always correct. Store reads would add async I/O and error paths
- * for a theoretical edge case that does not exist in practice.
- *
- * @param sessionId - Process-local UUID (for logging correlation).
- * @param ctx - V2ToolContext from the shared DI container.
- * @param apiKey - Anthropic API key for the child session's Claude model.
- * @param thisWorkrailSessionId - WorkRail session ID of the parent (becomes parentSessionId in child).
- * @param currentDepth - Spawn depth of the parent session (0 for root sessions).
- * @param maxDepth - Maximum allowed spawn depth. Blocks spawn when currentDepth >= maxDepth.
- * @param runWorkflowFn - Injected runWorkflow function (allows testing without real LLM calls).
- * @param schemas - Plain JSON Schema map from getSchemas().
- * @param emitter - Optional event emitter for structured lifecycle events.
- */
-/**
- * Factory for the `spawn_agent` tool, which lets a parent session delegate sub-tasks
- * to child WorkRail sessions.
- *
- * LIMITATION -- branchStrategy: Child sessions spawned by this tool always have
- * `branchStrategy: 'none'`. They operate in the parent's workspace (or the workspace
- * provided via the spawn_agent params) without their own isolated worktree or feature
- * branch. The child session writes directly to whatever workspace path it is given.
- *
- * WHY: spawn_agent constructs a WorkflowTrigger without a branchStrategy field, so the
- * child defaults to 'none'. Creating an isolated worktree for each child session would
- * require git credentials, disk allocation, and a branch-naming scheme per child --
- * overhead that is unnecessary for most coordinator/sub-agent patterns.
- *
- * Coordinators that need isolated child sessions (i.e. a child that fetches a branch,
- * makes changes, and opens a PR independently) should dispatch them via
- * `TriggerRouter.dispatch()` instead, which supports the full trigger configuration
- * including branchStrategy: 'worktree'.
- */
-export function makeSpawnAgentTool(
-  sessionId: string,
-  ctx: V2ToolContext,
-  apiKey: string,
-  thisWorkrailSessionId: string,
-  currentDepth: number,
-  maxDepth: number,
-  runWorkflowFn: typeof runWorkflow,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  schemas: Record<string, any>,
-  emitter?: DaemonEventEmitter,
-  abortRegistry?: AbortRegistry,
-): AgentTool {
-  return {
-    name: 'spawn_agent',
-    description:
-      'Spawn a child WorkRail session to handle a delegated sub-task. ' +
-      'Blocks until the child session completes, then returns the child\'s outcome and notes. ' +
-      'Use this when a step requires delegating a well-defined sub-task to a separate workflow. ' +
-      'IMPORTANT: The parent session\'s time limit (maxSessionMinutes) keeps ticking while the child runs. ' +
-      'Configure the parent with enough time to cover both its own work and the child\'s work. ' +
-      'Per-trigger limits (maxOutputTokens, maxTurns, maxSessionMinutes) are NOT inherited by child sessions spawned via spawn_agent -- each child uses its own trigger\'s agentConfig. ' +
-      'Returns: { childSessionId, outcome: "success"|"error"|"timeout", notes: string, artifacts?: readonly unknown[] }. ' +
-      'On success, artifacts contains the child session\'s final step artifacts if any were produced. ' +
-      'Check outcome before using notes -- on error/timeout, notes contains the error message.',
-    inputSchema: schemas['SpawnAgentParams'],
-    label: 'Spawn Agent',
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    execute: async (_toolCallId: string, params: any): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.workflowId !== 'string' || !params.workflowId) throw new Error('spawn_agent: workflowId must be a non-empty string');
-      if (typeof params.goal !== 'string' || !params.goal) throw new Error('spawn_agent: goal must be a non-empty string');
-      if (typeof params.workspacePath !== 'string' || !params.workspacePath) throw new Error('spawn_agent: workspacePath must be a non-empty string');
-      console.log(`[WorkflowRunner] Tool: spawn_agent sessionId=${sessionId} workflowId=${String(params.workflowId)} depth=${currentDepth}/${maxDepth}`);
-      emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'spawn_agent', summary: `${String(params.workflowId)} depth=${currentDepth}`, ...withWorkrailSession(thisWorkrailSessionId) });
-
-      // ---- Depth limit enforcement (synchronous, before any async work) ----
-      // WHY check before executeStartWorkflow: fail fast at the boundary. A depth error
-      // is a configuration issue, not a transient failure. No child session should be
-      // created at all when the depth limit is exceeded.
-      if (currentDepth >= maxDepth) {
-        const limitResult = {
-          childSessionId: null,
-          outcome: 'error' as const,
-          notes: `Max spawn depth exceeded (currentDepth=${currentDepth}, maxDepth=${maxDepth}). ` +
-            `Cannot spawn a child session from this depth. ` +
-            `Increase agentConfig.maxSubagentDepth if deeper delegation is intentional.`,
-        };
-        return {
-          content: [{ type: 'text', text: JSON.stringify(limitResult) }],
-          details: limitResult,
-        };
-      }
-
-      // ---- Pre-create child session (Candidate 2: _preAllocatedStartResponse pattern) ----
-      // WHY call executeStartWorkflow first: childSessionId is deterministic and the child
-      // is observable in the store from the moment execute() is called. If the process crashes
-      // after executeStartWorkflow but before runWorkflow, the zombie session is traceable
-      // via parentSessionId. Adapts the proven pattern from console-routes.ts.
-      const startInput = {
-        workflowId: String(params.workflowId),
-        workspacePath: String(params.workspacePath),
-        goal: String(params.goal),
-      };
-
-      const startResult = await executeStartWorkflow(
-        startInput,
-        ctx,
-        // WHY parentSessionId via internalContext: the public V2StartWorkflowInput schema
-        // stays unchanged. parentSessionId is extracted in buildInitialEvents() to populate
-        // session_created.data (typed, durable, DAG-queryable).
-        // is_autonomous: 'true' marks the child as a daemon-owned session.
-        { is_autonomous: 'true', workspacePath: String(params.workspacePath), parentSessionId: thisWorkrailSessionId },
-      );
-
-      if (startResult.isErr()) {
-        const errResult = {
-          childSessionId: null,
-          outcome: 'error' as const,
-          notes: `Failed to start child workflow: ${startResult.error.kind} -- ${JSON.stringify(startResult.error)}`,
-        };
-        return {
-          content: [{ type: 'text', text: JSON.stringify(errResult) }],
-          details: errResult,
-        };
-      }
-
-      // ---- Decode childSessionId from continueToken ----
-      // WHY token decode (not return from executeStartWorkflow): adding sessionId to
-      // V2StartWorkflowOutputSchema would be a public API change (GAP-7 territory).
-      // Token decode via the alias store is the correct in-process path.
-      // On decode failure: proceed with childSessionId = null (observable in logs).
-      let childSessionId: string | null = null;
-      const childContinueToken = startResult.value.response.continueToken ?? '';
-      if (childContinueToken) {
-        const decoded = await parseContinueTokenOrFail(
-          childContinueToken,
-          ctx.v2.tokenCodecPorts,
-          ctx.v2.tokenAliasStore,
-        );
-        if (decoded.isOk()) {
-          childSessionId = decoded.value.sessionId;
-        } else {
-          console.warn(
-            `[WorkflowRunner] spawn_agent: could not decode childSessionId from continueToken -- ` +
-            `childSessionId will be null in result. Reason: ${decoded.error.message}`,
-          );
-        }
-      }
-
-      // ---- Run child workflow (blocking until complete) ----
-      // WHY direct runWorkflow() call (not dispatch()): dispatch() is fire-and-forget and uses
-      // a global Semaphore. Calling it from inside a running session would deadlock.
-      // Direct await runWorkflow() is naturally blocking -- the parent's AgentLoop is paused
-      // inside execute() until the child completes (AgentLoop.execute() is sequential).
-      const childResult = await runWorkflowFn(
-        {
-          workflowId: String(params.workflowId),
-          goal: String(params.goal),
-          workspacePath: String(params.workspacePath),
-          context: params.context as Readonly<Record<string, unknown>> | undefined,
-          // WHY spawnDepth: child session constructs its own spawn_agent tool at depth+1.
-          // This is the mechanism by which depth limits propagate through the tree.
-          spawnDepth: currentDepth + 1,
-          // WHY parentSessionId: threads the parent link through runWorkflow -> executeStartWorkflow
-          // for context_set injection (alongside the session_created.data written above).
-          parentSessionId: thisWorkrailSessionId,
-          // WHY _preAllocatedStartResponse: the session is already created above.
-          // runWorkflow() MUST NOT call executeStartWorkflow() again (invariant).
-          _preAllocatedStartResponse: startResult.value.response,
-        },
-        ctx,
-        apiKey,
-        undefined, // daemonRegistry: child sessions are not registered (no isLive tracking needed)
-        emitter,
-        undefined, // steerRegistry: child sessions are not steerable by coordinator
-        abortRegistry, // WHY: thread abort registry so child sessions are abortable on SIGTERM
-      ) as ChildWorkflowRunResult;
-      // WHY cast to ChildWorkflowRunResult: runWorkflow() returns WorkflowRunResult (4 variants)
-      // for TriggerRouter compatibility, but structurally only produces success/error/timeout.
-      // delivery_failed is produced by TriggerRouter after a callbackUrl POST fails -- a
-      // trigger-layer concern that does not apply here (child sessions have no callbackUrl).
-      // The cast documents this architectural invariant; assertNever below catches any future
-      // violation at compile time if ChildWorkflowRunResult or runWorkflow() changes.
-
-      // ---- Map ChildWorkflowRunResult to structured output ----
-      // WHY ChildWorkflowRunResult (not WorkflowRunResult): runWorkflow() never produces
-      // delivery_failed -- see ChildWorkflowRunResult type definition and WHY comment above.
-      // Using the narrower type gives compile-time exhaustiveness over the 3 real variants;
-      // assertNever guards against future additions.
-      let resultObj: {
-        childSessionId: string | null;
-        outcome: 'success' | 'error' | 'timeout' | 'stuck';
-        notes: string;
-        artifacts?: readonly unknown[];
-        issueSummaries?: readonly string[];
-      };
-
-      if (childResult._tag === 'success') {
-        resultObj = {
-          childSessionId,
-          outcome: 'success',
-          notes: childResult.lastStepNotes ?? '(no notes from child session)',
-          // WHY spread conditional: artifacts must be absent (not null/undefined) when the child
-          // produced none. Mirrors the WorkflowRunSuccess construction pattern at lines 2868-2869.
-          ...(childResult.lastStepArtifacts !== undefined ? { artifacts: childResult.lastStepArtifacts } : {}),
-        };
-      } else if (childResult._tag === 'error') {
-        resultObj = {
-          childSessionId,
-          outcome: 'error',
-          notes: childResult.message,
-        };
-      } else if (childResult._tag === 'timeout') {
-        resultObj = {
-          childSessionId,
-          outcome: 'timeout',
-          notes: childResult.message,
-        };
-      } else if (childResult._tag === 'stuck') {
-        resultObj = {
-          childSessionId,
-          outcome: 'stuck',
-          notes: childResult.message,
-          ...(childResult.issueSummaries !== undefined
-            ? { issueSummaries: childResult.issueSummaries }
-            : {}),
-        };
-      } else {
-        // Compile-time exhaustiveness guard. If ChildWorkflowRunResult gains a new variant
-        // without a corresponding branch above, TypeScript will emit a compile error here.
-        // At runtime this is unreachable -- see ChildWorkflowRunResult and WHY comment above.
-        assertNever(childResult);
-      }
-
-      console.log(`[WorkflowRunner] spawn_agent completed: sessionId=${sessionId} childSessionId=${childSessionId ?? 'null'} outcome=${resultObj.outcome}`);
-      emitter?.emit({ kind: 'tool_called', sessionId, toolName: 'spawn_agent_complete', summary: `outcome=${resultObj.outcome} child=${childSessionId ?? 'null'}`, ...withWorkrailSession(thisWorkrailSessionId) });
-
-      return {
-        content: [{ type: 'text', text: JSON.stringify(resultObj) }],
-        details: resultObj,
-      };
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// report_issue tool
-// ---------------------------------------------------------------------------
-
-/**
- * Append a single JSON issue record to the per-session JSONL file.
- *
- * WHY void + catch: issue recording is purely observational. A failed write
- * (disk full, permission denied) must not propagate to the caller or interrupt
- * the workflow session. Same fire-and-forget contract as DaemonEventEmitter.
- *
- * WHY separate helper: keeps execute() synchronous from the caller's perspective
- * and makes the async write path independently testable via issuesDirOverride.
- *
- * @param issuesDir - Directory for issue files (override in tests; production uses ~/.workrail/issues).
- * @param sessionId - Session identifier used as the filename.
- * @param record - The issue payload to serialize as a JSON line.
- */
 /**
  * Append a stuck-escalation entry to ~/.workrail/outbox.jsonl.
  *
@@ -2997,313 +1830,6 @@ async function writeStuckOutboxEntry(opts: {
         `${err instanceof Error ? err.message : String(err)}`,
     );
   }
-}
-
-async function appendIssueAsync(
-  issuesDir: string,
-  sessionId: string,
-  record: IssueRecord,
-): Promise<void> {
-  await fs.mkdir(issuesDir, { recursive: true });
-  const filePath = path.join(issuesDir, `${sessionId}.jsonl`);
-  const line = JSON.stringify({ ...record, ts: Date.now() }) + '\n';
-  await fs.appendFile(filePath, line, 'utf8');
-}
-
-/** Input record shape for a report_issue call (before ts is appended). */
-interface IssueRecord {
-  sessionId: string;
-  kind: 'tool_failure' | 'blocked' | 'unexpected_behavior' | 'needs_human' | 'self_correction';
-  severity: 'info' | 'warn' | 'error' | 'fatal';
-  summary: string;
-  context?: string;
-  toolName?: string;
-  command?: string;
-  suggestedFix?: string;
-  continueToken?: string;
-}
-
-/**
- * Build the report_issue tool.
- *
- * Agents call this to record a structured issue for the auto-fix coordinator.
- * The tool does NOT stop the session -- it creates a record and returns a
- * confirmation. For fatal severity, the return value instructs the agent to call
- * continue_workflow with a blocker note, after which the session ends.
- *
- * @param sessionId - The process-local session UUID (keys the issues file).
- * @param emitter - Optional event emitter to fire an issue_reported event.
- * @param workrailSessionId - The WorkRail session ID for event correlation (optional).
- * @param issuesDirOverride - Override the issues directory (for tests).
- * @param onIssueSummary - Optional callback called synchronously with the issue summary
- *   string after each successful report_issue call. Used by runWorkflow() to accumulate
- *   issue summaries for the WORKTRAIN_STUCK marker without async file I/O.
- *   WHY optional callback: avoids circular dependency and keeps execute() synchronous
- *   from the caller's perspective. Fire-and-forget writes happen separately.
- */
-export function makeReportIssueTool(
-  sessionId: string,
-  emitter?: DaemonEventEmitter,
-  workrailSessionId?: string | null,
-  issuesDirOverride?: string,
-  onIssueSummary?: (summary: string) => void,
-): AgentTool {
-  const issuesDir = issuesDirOverride ?? path.join(os.homedir(), '.workrail', 'issues');
-
-  return {
-    name: 'report_issue',
-    description:
-      "Record a structured issue, error, or unexpected behavior. Call this AND continue_workflow (unless fatal). " +
-      "Does not stop the session -- it creates a record for the auto-fix coordinator.",
-    inputSchema: {
-      type: 'object',
-      properties: {
-        kind: {
-          type: 'string',
-          enum: ['tool_failure', 'blocked', 'unexpected_behavior', 'needs_human', 'self_correction'],
-          description: 'Category of issue being reported.',
-        },
-        severity: {
-          type: 'string',
-          enum: ['info', 'warn', 'error', 'fatal'],
-          description: 'Severity level. Fatal means the session cannot continue productively.',
-        },
-        summary: {
-          type: 'string',
-          description: 'One-line summary of the issue. Max 200 chars.',
-          maxLength: 200,
-        },
-        context: {
-          type: 'string',
-          description: 'What you were trying to do when this issue occurred.',
-        },
-        toolName: {
-          type: 'string',
-          description: 'Name of the tool that failed or behaved unexpectedly, if applicable.',
-        },
-        command: {
-          type: 'string',
-          description: 'The shell command or expression that caused the issue, if applicable.',
-        },
-        suggestedFix: {
-          type: 'string',
-          description: 'A suggested fix or recovery action for the auto-fix coordinator.',
-        },
-        continueToken: {
-          type: 'string',
-          description: 'The current continueToken, so the coordinator can resume this session.',
-        },
-      },
-      required: ['kind', 'severity', 'summary'],
-      additionalProperties: false,
-    },
-    label: 'report_issue',
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    execute: async (_toolCallId: string, params: any): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.kind !== 'string' || !params.kind) throw new Error('report_issue: kind must be a non-empty string');
-      if (typeof params.severity !== 'string' || !params.severity) throw new Error('report_issue: severity must be a non-empty string');
-      if (typeof params.summary !== 'string' || !params.summary) throw new Error('report_issue: summary must be a non-empty string');
-      const record: IssueRecord = {
-        sessionId,
-        kind: params.kind as IssueRecord['kind'],
-        severity: params.severity as IssueRecord['severity'],
-        summary: String(params.summary ?? '').slice(0, 200),
-        ...(params.context !== undefined && { context: String(params.context) }),
-        ...(params.toolName !== undefined && { toolName: String(params.toolName) }),
-        ...(params.command !== undefined && { command: String(params.command) }),
-        ...(params.suggestedFix !== undefined && { suggestedFix: String(params.suggestedFix) }),
-        ...(params.continueToken !== undefined && { continueToken: String(params.continueToken) }),
-      };
-
-      // Fire-and-forget: write must never block execute() or propagate errors.
-      // WHY void + catch: observability must not affect correctness.
-      void appendIssueAsync(issuesDir, sessionId, record).catch(() => {
-        // Intentionally empty: write failures are silently swallowed.
-      });
-
-      // Emit structured event for console/SSE stream visibility.
-      emitter?.emit({
-        kind: 'issue_reported',
-        sessionId,
-        issueKind: record.kind,
-        severity: record.severity,
-        summary: record.summary,
-        ...(record.continueToken !== undefined && { continueToken: record.continueToken }),
-        ...(workrailSessionId != null ? { workrailSessionId } : {}),
-      });
-
-      // Notify the accumulator so runWorkflow() can include issue summaries in
-      // the WORKTRAIN_STUCK marker without async file I/O.
-      // WHY synchronous callback: execute() already runs synchronously from the
-      // agent loop's perspective; the callback push is O(1) and never throws.
-      onIssueSummary?.(record.summary);
-
-      const isFatal = record.severity === 'fatal';
-      const message = isFatal
-        ? `FATAL issue recorded. Call continue_workflow with notes explaining the blocker, then the session will end.`
-        : `Issue recorded (severity=${record.severity}). Continue with your work unless this is fatal.`;
-
-      return {
-        content: [{ type: 'text', text: message }],
-        details: { sessionId, kind: record.kind, severity: record.severity },
-      };
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// signal_coordinator tool
-// ---------------------------------------------------------------------------
-
-/**
- * Directory that holds per-session signal JSONL files.
- *
- * Each concurrent runWorkflow() call appends its signals to
- * ~/.workrail/signals/<sessionId>.jsonl. Coordinators can tail this file
- * or read it at session boundaries without touching the durable session store.
- *
- * WHY a sidecar file instead of the v2 session store:
- * The session store uses a re-entrancy guard (ExecutionSessionGate.activeSessions).
- * At the point signal_coordinator executes, the gate is already held by the
- * ongoing continue_workflow / complete_step machinery. Attempting a second
- * session lock would return SESSION_LOCK_REENTRANT. The sidecar file avoids
- * this entirely: it is a separate, lock-free append channel that the coordinator
- * can read independently. The DaemonEventEmitter simultaneously broadcasts the
- * signal to the daemon JSONL event stream for live console visibility.
- */
-export const DAEMON_SIGNALS_DIR = path.join(os.homedir(), '.workrail', 'signals');
-
-/**
- * Append a single JSON signal record to the per-session JSONL file.
- *
- * Fire-and-forget: errors are swallowed so a failed write never interrupts
- * the session. Same contract as appendIssueAsync and DaemonEventEmitter.
- */
-async function appendSignalAsync(
-  signalsDir: string,
-  sessionId: string,
-  record: SignalRecord,
-): Promise<void> {
-  await fs.mkdir(signalsDir, { recursive: true });
-  const filePath = path.join(signalsDir, `${sessionId}.jsonl`);
-  const line = JSON.stringify({ ...record, ts: Date.now() }) + '\n';
-  await fs.appendFile(filePath, line, 'utf8');
-}
-
-/** Payload written to the per-session JSONL sidecar (before ts is appended). */
-interface SignalRecord {
-  readonly signalId: string;
-  readonly sessionId: string;
-  readonly workrailSessionId?: string;
-  readonly signalKind: string;
-  readonly payload: Readonly<Record<string, unknown>>;
-}
-
-/**
- * Build the signal_coordinator tool for daemon sessions.
- *
- * The agent calls this to emit a structured coordinator signal without
- * advancing the workflow step. The signal is written to:
- * 1. ~/.workrail/signals/<sessionId>.jsonl -- sidecar JSONL for coordinator polling
- * 2. The daemon event log (via DaemonEventEmitter) -- for live console visibility
- *
- * WHY NOT writing to the v2 session store directly:
- * The session store uses a re-entrancy guard (ExecutionSessionGate). When
- * signal_coordinator executes, the gate is already held by the in-flight
- * continue_workflow / complete_step tool. Attempting withHealthySessionLock()
- * would return SESSION_LOCK_REENTRANT and abort the signal write. The sidecar
- * file is the correct channel for mid-step, non-advancing signal emission.
- *
- * WHY fire-and-observe (always returns immediately):
- * The tool must not block the agent. Signals are best-effort observability
- * artifacts -- a coordinator that reads them asynchronously is the intended
- * consumer. The agent proceeds to its next tool call regardless.
- *
- * @param sessionId - Process-local UUID (keys the sidecar JSONL file).
- * @param emitter - Optional event emitter for daemon JSONL visibility.
- * @param workrailSessionId - WorkRail session ID for event correlation.
- * @param signalsDirOverride - Override the signals directory (for tests).
- */
-export function makeSignalCoordinatorTool(
-  sessionId: string,
-  emitter?: DaemonEventEmitter,
-  workrailSessionId?: string | null,
-  signalsDirOverride?: string,
-): AgentTool {
-  const signalsDir = signalsDirOverride ?? DAEMON_SIGNALS_DIR;
-
-  return {
-    name: 'signal_coordinator',
-    description:
-      'Emit a structured mid-session signal to the coordinator WITHOUT advancing the workflow step. ' +
-      'Use this to surface progress updates, intermediate findings, data requests, ' +
-      'approval requests, or blocking conditions while the session continues. ' +
-      'Always returns immediately -- fire-and-observe, never blocks. ' +
-      'Signal kinds: "progress" (heartbeat, no data needed), "finding" (intermediate result), ' +
-      '"data_needed" (request external data), "approval_needed" (request coordinator approval), ' +
-      '"blocked" (cannot continue without coordinator intervention).',
-    inputSchema: {
-      type: 'object',
-      properties: {
-        signalKind: {
-          type: 'string',
-          enum: ['progress', 'finding', 'data_needed', 'approval_needed', 'blocked'],
-          description: 'The kind of signal to emit.',
-        },
-        payload: {
-          type: 'object',
-          additionalProperties: true,
-          description: 'Structured data accompanying the signal. Pass {} for progress signals.',
-        },
-      },
-      required: ['signalKind', 'payload'],
-      additionalProperties: false,
-    },
-    label: 'signal_coordinator',
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    execute: async (_toolCallId: string, params: any): Promise<AgentToolResult<unknown>> => {
-      if (typeof params.signalKind !== 'string' || !params.signalKind) throw new Error('signal_coordinator: signalKind must be a non-empty string');
-      const signalId = 'sig_' + randomUUID().replace(/-/g, '').slice(0, 8);
-      const signalKind = String(params.signalKind ?? 'progress');
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const payload = (typeof params.payload === 'object' && params.payload !== null && !Array.isArray(params.payload))
-        ? (params.payload as Record<string, unknown>)
-        : {};
-
-      console.log(`[WorkflowRunner] Tool: signal_coordinator sessionId=${sessionId} signalKind=${signalKind} signalId=${signalId}`);
-
-      const record: SignalRecord = {
-        signalId,
-        sessionId,
-        ...(workrailSessionId != null ? { workrailSessionId } : {}),
-        signalKind,
-        payload,
-      };
-
-      // Fire-and-forget sidecar write. A failed write never blocks or throws.
-      void appendSignalAsync(signalsDir, sessionId, record).catch(() => {
-        // Intentionally empty: write failures are silently swallowed.
-      });
-
-      // Emit to the daemon event log for live console visibility.
-      emitter?.emit({
-        kind: 'signal_emitted',
-        sessionId,
-        signalKind,
-        signalId,
-        payload,
-        ...(workrailSessionId != null ? { workrailSessionId } : {}),
-      });
-
-      const result = { status: 'recorded' as const, signalId };
-      return {
-        content: [{ type: 'text', text: JSON.stringify(result) }],
-        details: result,
-      };
-    },
-  };
 }
 
 // ---------------------------------------------------------------------------
